backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.1",
    "httpx>=0.28.1",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.40.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
